+++ firebase-ml-modeldownloader/src/test/java/com/google/firebase/ml/modeldownloader/internal/CustomModelDownloadServiceTest.java
@@ -121,6 +121,20 @@ public class CustomModelDownloadServiceTest {
   private FirebaseInstallationsApi installationsApiMock;
   @Mock private FirebaseMlLogger mockEventLogger;
 
+  private final ModelFileDownloadService modelFileDownloadService =
+      mock(ModelFileDownloadService.class);
+  private final CustomModel.Factory modelFactory =
+      (name, modelHash, fileSize, downloadId, localFilePath, downloadUrl, downloadUrlExpiry) ->
+          new CustomModel(
+              modelFileDownloadService,
+              name,
+              modelHash,
+              fileSize,
+              downloadId,
+              localFilePath,
+              downloadUrl,
+              downloadUrlExpiry);
+
   @Before
   public void setUp() {
     MockitoAnnotations.initMocks(this);
@@ -172,18 +186,20 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getNewDownloadUrlWithExpiry(PROJECT_ID, MODEL_NAME);
 
     Assert.assertEquals(
         modelTask.getResult(),
-        new CustomModel(MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
+        modelFactory.create(
+            MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
 
     WireMock.verify(
         getRequestedFor(urlEqualTo(downloadPath))
@@ -220,18 +236,20 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             null,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getNewDownloadUrlWithExpiry(PROJECT_ID, MODEL_NAME);
 
     Assert.assertEquals(
         modelTask.getResult(),
-        new CustomModel(MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
+        modelFactory.create(
+            MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
 
     WireMock.verify(
         getRequestedFor(urlEqualTo(downloadPath))
@@ -266,18 +284,20 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
     Assert.assertEquals(
         modelTask.getResult(),
-        new CustomModel(MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
+        modelFactory.create(
+            MODEL_NAME, MODEL_HASH, FILE_SIZE, DOWNLOAD_URI, TEST_EXPIRATION_IN_MS));
 
     WireMock.verify(
         getRequestedFor(urlEqualTo(downloadPath))
@@ -314,12 +334,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -363,12 +384,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -415,12 +437,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -473,12 +496,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -531,12 +555,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -588,12 +613,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -643,12 +669,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -690,12 +717,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -718,12 +746,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, null);
 
@@ -747,12 +776,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             "https7://localhost:8989/barUrl",
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);
 
@@ -775,12 +805,13 @@ public class CustomModelDownloadServiceTest {
     CustomModelDownloadService service =
         new CustomModelDownloadService(
             ApplicationProvider.getApplicationContext(),
-            installationsApiMock,
+            () -> installationsApiMock,
             directExecutor,
             API_KEY,
             PACKAGE_FINGERPRINT_HASH,
             TEST_ENDPOINT,
-            mockEventLogger);
+            mockEventLogger,
+            modelFactory);
 
     Task<CustomModel> modelTask = service.getCustomModelDetails(PROJECT_ID, MODEL_NAME, MODEL_HASH);